refactor: move `WasmPermissions` to dedicated module
